+++ blaze-client/src/main/scala/org/http4s/client/blaze/BlazeHttp1ClientParser.scala
@@ -46,7 +46,7 @@ final private class BlazeHttp1ClientParser extends Http1ClientParser {
                                             scheme: String,
                                             majorversion: Int,
                                             minorversion: Int): Unit = {
-    status = Status.fromIntAndReason(code, reason).valueOr(e => throw new ParseException(e))
+    status = Status.fromIntAndReason(code, reason).valueOr(throw _)
     httpVersion = {
       if (majorversion == 1 && minorversion == 1)  HttpVersion.`HTTP/1.1`
       else if (majorversion == 1 && minorversion == 0)  HttpVersion.`HTTP/1.0`