+++ blaze-core/src/test/scala/org/http4s/blaze/util/FailingWriter.scala
@@ -1,0 +1,17 @@
+package org.http4s.blaze.util
+
+import org.http4s.blaze.pipeline.Command.EOF
+import scodec.bits.ByteVector
+
+import scala.concurrent.{ExecutionContext, Future}
+
+
+
+class FailingWriter() extends ProcessWriter {
+
+  override implicit protected def ec: ExecutionContext = scala.concurrent.ExecutionContext.global
+
+  override protected def writeEnd(chunk: ByteVector): Future[Boolean] = Future.failed(EOF)
+
+  override protected def writeBodyChunk(chunk: ByteVector, flush: Boolean): Future[Unit] = Future.failed(EOF)
+}